AC-816 fixed and update code of AC-795

+++ AppCMS/src/main/java/com/viewlift/models/data/appcms/api/Statistics.java
@@ -13,7 +13,7 @@ public class Statistics {
 
     @SerializedName("tmdbRatingAvg")
     @Expose
-    int tmdbRatingAvg;
+    double tmdbRatingAvg;
 
     @SerializedName("averageViewerGrade")
     @Expose
@@ -27,7 +27,7 @@ public class Statistics {
         this.tmdbRatingCount = tmdbRatingCount;
     }
 
-    public int getTmdbRatingAvg() {
+    public double getTmdbRatingAvg() {
         return tmdbRatingAvg;
     }
